rename class; use util methods for serialization and deserialization
