+++ app/src/org/commcare/UpdateToPrompt.java
@@ -1,12 +1,16 @@
 package org.commcare;
 
+import android.app.Activity;
 import android.content.Context;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.util.Base64;
 
+import org.commcare.activities.PromptUpdateActivity;
 import org.commcare.logging.AndroidLogger;
+import org.commcare.utils.SerializationUtil;
 import org.javarosa.core.model.utils.DateUtils;
 import org.javarosa.core.services.Logger;
 import org.javarosa.core.util.externalizable.DeserializationException;
@@ -15,7 +19,7 @@ import org.javarosa.core.util.externalizable.ExtWrapNullable;
 import org.javarosa.core.util.externalizable.Externalizable;
 import org.javarosa.core.util.externalizable.PrototypeFactory;
 
-import java.io.ByteArrayOutputStream;
+import java.io.ByteArrayInputStream;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -69,7 +73,7 @@ public class UpdateToPrompt implements Externalizable {
         } else {
             // If the latest signal we're getting is that our current version is up-to-date,
             // then we should wipe any update prompt for this type that was previously stored
-            CommCareHeartbeatManager.wipeStoredUpdate(this.isApkUpdate);
+            wipeStoredUpdate(this.isApkUpdate);
         }
     }
 
@@ -106,35 +110,70 @@ public class UpdateToPrompt implements Externalizable {
     }
 
     private void writeToPrefsObject(SharedPreferences prefs) {
-        ByteArrayOutputStream baos = null;
-        DataOutputStream serializedStream = null;
         try {
-            baos = new ByteArrayOutputStream();
-            serializedStream = new DataOutputStream(baos);
-            try {
-                ExtUtil.write(serializedStream, this);
-                String serializedString = Base64.encodeToString(baos.toByteArray(), Base64.DEFAULT);
-                prefs.edit().putString(
-                        isApkUpdate ? KEY_APK_UPDATE_TO_PROMPT : KEY_CCZ_UPDATE_TO_PROMPT,
-                        serializedString).commit();
-                baos.close();
-                serializedStream.close();
-            } catch (IOException e) {
-                Logger.log(AndroidLogger.TYPE_ERROR_WORKFLOW,
-                        "IO error encountered while serializing UpdateToPrompt: " + e.getMessage());
-            }
-        } finally {
-            try {
-                if (baos != null) {
-                    baos.close();
+            byte[] serializedBytes = SerializationUtil.serialize(this);
+            String serializedString = Base64.encodeToString(serializedBytes, Base64.DEFAULT);
+            prefs.edit().putString(
+                    isApkUpdate ? KEY_APK_UPDATE_TO_PROMPT : KEY_CCZ_UPDATE_TO_PROMPT,
+                    serializedString).commit();
+        } catch (Exception e) {
+            Logger.log(AndroidLogger.TYPE_ERROR_WORKFLOW,
+                    "Error encountered while serializing UpdateToPrompt: " + e.getMessage());
+        }
+    }
+
+    /**
+     * @return - If the user was prompted to update
+     */
+    public static boolean promptForUpdateIfNeeded(Activity context) {
+        UpdateToPrompt cczUpdate = getCurrentUpdateToPrompt(false);
+        UpdateToPrompt apkUpdate = getCurrentUpdateToPrompt(true);
+        if (cczUpdate != null || apkUpdate != null) {
+            Intent i = new Intent(context, PromptUpdateActivity.class);
+            context.startActivity(i);
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * @return an UpdateToPrompt that has been stored in SharedPreferences and is still relevant
+     * (i.e. the user hasn't updated to or past this version since we stored it)
+     */
+    public static UpdateToPrompt getCurrentUpdateToPrompt(boolean forApkUpdate) {
+        CommCareApp currentApp = CommCareApplication.instance().getCurrentApp();
+        if (currentApp != null) {
+            String prefsKey = forApkUpdate ?
+                    UpdateToPrompt.KEY_APK_UPDATE_TO_PROMPT : UpdateToPrompt.KEY_CCZ_UPDATE_TO_PROMPT;
+            String serializedUpdate = currentApp.getAppPreferences().getString(prefsKey, "");
+            if (!"".equals(serializedUpdate)) {
+                try {
+                    byte[] updateBytes = Base64.decode(serializedUpdate, Base64.DEFAULT);
+                    UpdateToPrompt update = SerializationUtil.deserialize(updateBytes, UpdateToPrompt.class);
+                    if (update.isNewerThanCurrentVersion(currentApp)) {
+                        return update;
+                    } else {
+                        // The update we had stored is no longer relevant, so wipe it and return nothing
+                        wipeStoredUpdate(forApkUpdate);
+                        return null;
+                    }
+                } catch (Exception e) {
+                    // Something went wrong, so clear out whatever is there
+                    Logger.log(AndroidLogger.TYPE_ERROR_WORKFLOW,
+                            "IO error encountered while de-serializing saved UpdateToPrompt: "
+                                    + e.getMessage());
+                    wipeStoredUpdate(forApkUpdate);
                 }
-                if (serializedStream != null) {
-                    serializedStream.close();
-                }
-            } catch (IOException e) {
-                // Nothing we can do
             }
         }
+        return null;
+    }
+
+    public static void wipeStoredUpdate(boolean forApkUpdate) {
+        String prefsKey = forApkUpdate ?
+                UpdateToPrompt.KEY_APK_UPDATE_TO_PROMPT : UpdateToPrompt.KEY_CCZ_UPDATE_TO_PROMPT;
+        CommCareApplication.instance().getCurrentApp().getAppPreferences().edit()
+                .putString(prefsKey, "").commit();
     }
 
     @Override
