+++ app/src/org/commcare/activities/HomeScreenBaseActivity.java
@@ -16,7 +16,8 @@ import android.widget.AdapterView;
 import android.widget.Toast;
 
 import org.commcare.CommCareApplication;
-import org.commcare.CommCareHeartbeatManager;
+import org.commcare.HeartbeatLifecycleManager;
+import org.commcare.UpdateToPrompt;
 import org.commcare.activities.components.FormEntryConstants;
 import org.commcare.activities.components.FormEntryInstanceState;
 import org.commcare.activities.components.FormEntrySessionWrapper;
@@ -72,7 +73,6 @@ import org.javarosa.core.model.instance.TreeReference;
 import org.javarosa.core.services.Logger;
 import org.javarosa.core.services.locale.Localization;
 import org.javarosa.xpath.XPathTypeMismatchException;
-import org.json.JSONException;
 
 import java.io.File;
 import java.text.SimpleDateFormat;
@@ -193,7 +193,7 @@ public abstract class HomeScreenBaseActivity<T> extends SyncCapableCommCareActiv
                 showDemoModeWarning();
                 return;
             }
-            if (CommCareHeartbeatManager.promptForUpdateIfNeeded(this)) {
+            if (UpdateToPrompt.promptForUpdateIfNeeded(this)) {
                 return;
             }
             if (checkForPinLaunchConditions()) {
